Adding profiles to dao/stub
