+++ src/test/java/com/enterprise4045/groceryplanner/GroceryPlannerApplicationTests.java
@@ -1,7 +1,6 @@
 package com.enterprise4045.groceryplanner;
 
 import com.enterprise4045.groceryplanner.dao.ILoggedItemDAO;
-import com.enterprise4045.groceryplanner.dao.LoggedItemDAOStub;
 import com.enterprise4045.groceryplanner.dto.LoggedItem;
 import com.enterprise4045.groceryplanner.service.LoggedItemServiceStub;
 import org.junit.jupiter.api.Test;
